debug macos

+++ src/integer64.c
@@ -1062,7 +1062,7 @@ __attribute__((no_sanitize("signed-integer-overflow"))) SEXP matmult_integer64_i
         warning("third value is %lld - %d\n", tempsum,naflag);
         if(((cumsum) > 0) ? ((addValue) < (tempsum)) : ! ((addValue) < (tempsum))) warning("GOODISUM64 is true");
         if(((cumsum) > 0)) warning("(cumsum) > 0 is true");
-        if(((addValue) < (tempsum))) warning("(addValue) < (tempsum) is true");
+        if((((long double) (addValue)) < ((long double) (tempsum)))) warning("(addValue) < (tempsum) is true");
         if(!GOODISUM64(cumsum, addValue, tempsum)){
           warning("GOODISUM64 is false");
           naflag = TRUE;
